requirements.txt
streamlit
watchdog
openai
pyyaml
